+++ app/src/main/java/com/example/fridgetracker/fragments/PantryFragment.kt
@@ -97,13 +97,13 @@ class PantryFragment : Fragment() {
         mAlertDialog.addFoodButton.setOnClickListener {
             val foodName = dialogView.foodNameEntered.text.toString()
             val foodDate = dialogView.foodDateEntered.text.toString()
-            val foodQuantity = dialogView.foodQuantityEntered.text.toString().toInt()
+            val foodQuantity = dialogView.foodQuantityEntered.text.toString()
             val foodNote = dialogView.foodNoteEntered.text.toString()
 
             //store food into Food
             // If the string is empty, we do not want to accept that as an input
             if(foodName != "" && foodDate.toString() != "" && foodQuantity.toString() != "" && foodNote != ""){
-                val food = Food("pantry",foodName,foodDate,foodQuantity,foodNote)
+                val food = Food("pantry",foodName,foodDate,foodQuantity.toInt(),foodNote)
                 viewModel!!.insertFood(food)
                 mAlertDialog.dismiss()
             } else {
